rename: StringCalculator to TextCalculator
